// lib/generated/NetworkRoute.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { StopSpacing } from "./StopSpacing";

export type NetworkRoute = { name: string, mode: string, hourlyDepartures: Array<number>, headwayCv: number | null, peakHeadwayCv: number | null, tortuosity: number | null, straightLineKm: number | null, stopSpacing: StopSpacing | null, peakDepPerHour: number, baseDepPerHour: number, peakToBaseRatio: number | null, };