// lib/generated/Fleet.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { InterlinedBlock } from "./InterlinedBlock";

export type Fleet = { tramBlocks: number, busBlocks: number, railBlocks: number, totalBlocks: number, interlinedBlocks: number, interlinedExamples: Array<InterlinedBlock>, };