// lib/generated/RouteStatsOutput.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { MultimodalConnections } from "./MultimodalConnections";
import type { RouteStatsRoute } from "./RouteStatsRoute";
import type { RouteStatsSummary } from "./RouteStatsSummary";
import type { TransferHub } from "./TransferHub";

export type RouteStatsOutput = { generatedAt: string, summary: RouteStatsSummary, routes: Array<RouteStatsRoute>, transferHubs: Array<TransferHub>, multimodalConnections: MultimodalConnections, };